requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
chromadb>=0.4.15
openai>=1.3.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
streamlit-chat>=0.1.1